move remaining enums to enums.py
