+++ osrlib/osrlib/enums.py
@@ -1,8 +1,39 @@
 from enum import Enum
 
-class CharacterClassType(Enum):
-    """Specifies the class, or profession, of a player character or NPC."""
 
+class AbilityType(Enum):
+    STRENGTH = "Strength"
+    INTELLIGENCE = "Intelligence"
+    WISDOM = "Wisdom"
+    DEXTERITY = "Dexterity"
+    CONSTITUTION = "Constitution"
+    CHARISMA = "Charisma"
+
+
+class AttackType(Enum):
+    MELEE = "Meelee"
+    RANGED = "Ranged"
+    DEATH_RAY_POISON = "Death ray or poison"
+    MAGIC_WANDS = "Magic wands"
+    PARALYSIS_TURN_TO_STONE = "Paralysis or turn to stone"
+    DRAGON_BREATH = "Dragon breath"
+    RODS_STAVES_SPELLS = "Rods, staves, or spells"
+
+
+class ModifierType(Enum):
+    TO_HIT = "To hit"
+    DAMAGE = "Damage"
+    OPEN_DOORS = "Open doors"
+    LANGUAGES = "Languages"
+    SAVING_THROWS = "Magic-based saving throws"
+    AC = "AC"
+    INITIATIVE = "Initiative"
+    HP = "HP"
+    REACTION = "Monster and NPC reactions"
+    XP = "XP"
+
+
+class CharacterClassType(Enum):
     CLERIC = "Cleric"
     DWARF = "Dwarf"
     ELF = "Elf"
@@ -13,8 +44,19 @@ class CharacterClassType(Enum):
     COMMONER = "Commoner"
 
 
+class ItemType(Enum):
+    ARMOR = ("Armor", "Armor, helmet, gloves, or boots")
+    WEAPON = ("Weapon", "Bladed, blunt, or ranged weapon")
+    SPELL = ("Spell", "Spell or scroll")
+    EQUIPMENT = ("Equipment", "Piece of adventurers' equipment")
+    MAGIC_ITEM = (
+        "Magic item",
+        "Potion, ring, or other item imbued with magical properties",
+    )
+    ITEM = ("Normal item", "Normal (non-magical) item")
+
+
 class OpenAIModelVersion(Enum):
-    """Specifies the version of the OpenAI model to use."""
     DEFAULT = "gpt-3.5-turbo-1106"
     GPT35 = "gpt-3.5"
     GPT35TURBO = "gpt-3.5-turbo-1106"
@@ -23,9 +65,37 @@ class OpenAIModelVersion(Enum):
 
 
 class PartyReaction(Enum):
-    """Specifies the reaction of a PC or monster party at the start of an encounter."""
-
     FIGHT = "Fight"
     RUN = "Run"
     TALK = "Talk"
     PASS = "Pass"
+
+
+class TreasureType(Enum):
+    NONE = "None"
+    A = "A"
+    B = "B"
+    C = "C"
+    D = "D"
+    E = "E"
+    F = "F"
+    G = "G"
+    H = "H"
+    I = "I"
+    J = "J"
+    K = "K"
+    L = "L"
+    M = "M"
+    N = "N"
+    O = "O"
+    P = "P"
+    Q = "Q"
+    R = "R"
+    S = "S"
+    T = "T"
+    U = "U"
+    V = "V"
+    W = "W"
+    X = "X"
+    Y = "Y"
+    Z = "Z"
