+++ osrlib/osrlib/monster.py
@@ -1,8 +1,7 @@
 from typing import List
 from osrlib.dice_roller import roll_dice, DiceRoll
-from osrlib.enums import CharacterClassType
+from osrlib.enums import CharacterClassType, TreasureType
 from osrlib.player_character import Alignment
-from osrlib.treasure import TreasureType
 from osrlib.saving_throws import get_saving_throws_for_class_and_level
 from osrlib.game_manager import logger
