+++ osrlib/osrlib/treasure.py
@@ -1,37 +1,6 @@
-from enum import Enum
 from osrlib.dice_roller import roll_dice
-from osrlib.item import Item, ItemType, Weapon, Armor
-
-class TreasureType(Enum):
-    """Represents the treasure type of a treasure."""
-
-    NONE = "None"
-    A = "A"
-    B = "B"
-    C = "C"
-    D = "D"
-    E = "E"
-    F = "F"
-    G = "G"
-    H = "H"
-    I = "I"
-    J = "J"
-    K = "K"
-    L = "L"
-    M = "M"
-    N = "N"
-    O = "O"
-    P = "P"
-    Q = "Q"
-    R = "R"
-    S = "S"
-    T = "T"
-    U = "U"
-    V = "V"
-    W = "W"
-    X = "X"
-    Y = "Y"
-    Z = "Z"
+from osrlib.enums import ItemType, TreasureType
+from osrlib.item import Item, Weapon, Armor
 
 def get_treasure(treasure_type: TreasureType):
         """Get a collection of items appropriate for the specified treasure type.
