+++ osrlib/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "osrlib"
-version = "0.1.46"
+version = "0.1.47"
 description = "Turn-based dungeon-crawler game engine for OSR-style RPGs."
 authors = ["Marsh Macy <mmacy@users.noreply.github.com>"]
 license = "MIT"
