+++ tests/test_unit_armor.py
@@ -2,9 +2,7 @@ import logging
 
 import pytest
 
-from osrlib.ability import AbilityType
-from osrlib.character_classes import CharacterClassType
-from osrlib.combat import ModifierType
+from osrlib.enums import AbilityType, CharacterClassType, ModifierType
 from osrlib.item import Armor
 from osrlib.player_character import PlayerCharacter
